requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
langchain-openai==0.0.5
langchain-community
langchain
chromadb==0.4.22
pypdf
tiktoken